Updated the system to PHP 7.4 && changed from mysql to mysqli

+++ deleteinventory.php
@@ -1,13 +1,22 @@
 <?php
 
 // This is a sample code in case you wish to check the username from a mysql db table
-include('config/db.php');
+$dbhost = "localhost";
+$dbuser = "root";
+$dbpass = "";
+$dbname = "railway";
+$conn = mysqli_connect($dbhost, $dbuser, $dbpass, $dbname);
+if (!$conn){
+    die ('Could not connect: '.mysqli_error());
+}
+
+
 if($_GET['id'])
 {
 $id=$_GET['id'];
  $sql = "delete from reserve where id='$id'";
- mysql_query("delete from customer where id='$id'");
- mysql_query( $sql);
+ mysqli_query($conn,"delete from customer where id='$id'");
+ mysqli_query($conn, $sql);
 }
-
+mysqli_close($conn);
 ?>
